# JIT compile NTT