- JIT compile the NTT algorithm
- Add the radix-2 Cooley-Tukey algorithm