@mhostetter 2022-04-06
Just JIT compiling the `N^2` NTT, resulted in this minor speed-up.

### Before

```python
In [4]: %timeit galois.ntt([1, 2, 3, 4])
186 µs ± 59.9 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```

### After

```python
In [4]: %timeit galois.ntt([1, 2, 3, 4])
147 µs ± 1.37 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```