@mhostetter 2022-04-07
After adding the radix-2 Cooley-Tukey algorithm, the NTT performance improves like this:

### Before

```python
In [1]: import galois

In [2]: x = list(range(1, 128 + 1))

In [3]: galois.ntt(x)
Out[3]: 
GF([ 32,  16,  53,   3,  79, 114,  25, 159, 100, 213, 178,  28, 158, 119,
    146,  63, 206,  99, 254,  27, 246,  18, 116,  19, 128,  68, 170, 192,
    156,  70, 219,  67, 189, 148,  75,   9, 137, 177, 237,  14,  86, 187,
     96, 212, 203, 196, 109,  78, 214,  11, 220, 165,  31, 182,  72,  91,
     74, 253,  95, 144,   8,  82,  24,  39, 193,  90, 105,  47, 121, 242,
     34, 133,  55,  38,  57, 204,  98, 221, 166, 118, 172,  51,  20, 190,
    183, 174,  33, 199,  43, 115, 149, 209, 249, 120,  54, 238, 197,  62,
    167,  59, 230, 194, 216,  61,   1, 110,  13, 111, 140, 102, 132,  30,
    180,  66, 240,  10, 228, 101, 208, 173,  29, 227, 104,  15,  50, 126,
     76, 113], order=257)

In [4]: %timeit galois.ntt(x)
371 µs ± 1.6 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [5]: x = list(range(1, 2048 + 1))

In [6]: galois.ntt(x)
Out[6]: GF([ 9046,  7688, 11472, ...,  6532, 11058,  2553], order=12289)

In [7]: %timeit galois.ntt(x)
50.2 ms ± 118 µs per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### After

```python
In [1]: import galois

In [2]: x = list(range(1, 128 + 1))

In [3]: galois.ntt(x)
Out[3]: 
GF([ 32,  16,  53,   3,  79, 114,  25, 159, 100, 213, 178,  28, 158, 119,
    146,  63, 206,  99, 254,  27, 246,  18, 116,  19, 128,  68, 170, 192,
    156,  70, 219,  67, 189, 148,  75,   9, 137, 177, 237,  14,  86, 187,
     96, 212, 203, 196, 109,  78, 214,  11, 220, 165,  31, 182,  72,  91,
     74, 253,  95, 144,   8,  82,  24,  39, 193,  90, 105,  47, 121, 242,
     34, 133,  55,  38,  57, 204,  98, 221, 166, 118, 172,  51,  20, 190,
    183, 174,  33, 199,  43, 115, 149, 209, 249, 120,  54, 238, 197,  62,
    167,  59, 230, 194, 216,  61,   1, 110,  13, 111, 140, 102, 132,  30,
    180,  66, 240,  10, 228, 101, 208, 173,  29, 227, 104,  15,  50, 126,
     76, 113], order=257)

In [4]: %timeit galois.ntt(x)
227 µs ± 5.01 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)

In [5]: x = list(range(1, 2048 + 1))

In [6]: galois.ntt(x)
Out[6]: GF([ 9046,  7688, 11472, ...,  6532, 11058,  2553], order=12289)

In [7]: %timeit galois.ntt(x)
1.04 ms ± 6.46 µs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
```